requirements.txt
pandas>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
plotly>=5.10.0
requests>=2.28.0
python-dotenv>=1.0.0
seaborn>=0.11.0
pytest>=7.0.0
black>=23.0.0
isort>=5.10.0
sphinx>=5.0.0